chore(sequencer_infra): move server loop function
commit-id:c6a69640
